+++ test/mocks/CompTimelock.sol
@@ -4,43 +4,43 @@ pragma solidity ^0.8.18;
 contract CompTimelock {
     event NewAdmin(address indexed newAdmin);
     event NewPendingAdmin(address indexed newPendingAdmin);
-    event NewDelay(uint indexed newDelay);
+    event NewDelay(uint256 indexed newDelay);
     event CancelTransaction(
         bytes32 indexed txHash,
         address indexed target,
-        uint value,
+        uint256 value,
         string signature,
         bytes data,
-        uint eta
+        uint256 eta
     );
     event ExecuteTransaction(
         bytes32 indexed txHash,
         address indexed target,
-        uint value,
+        uint256 value,
         string signature,
         bytes data,
-        uint eta
+        uint256 eta
     );
     event QueueTransaction(
         bytes32 indexed txHash,
         address indexed target,
-        uint value,
+        uint256 value,
         string signature,
         bytes data,
-        uint eta
+        uint256 eta
     );
 
-    uint public constant GRACE_PERIOD = 14 days;
-    uint public constant MINIMUM_DELAY = 10;
-    uint public constant MAXIMUM_DELAY = 30 days;
+    uint256 public constant GRACE_PERIOD = 14 days;
+    uint256 public constant MINIMUM_DELAY = 10;
+    uint256 public constant MAXIMUM_DELAY = 30 days;
 
     address public admin;
     address public pendingAdmin;
-    uint public delay;
+    uint256 public delay;
 
     mapping(bytes32 => bool) public queuedTransactions;
 
-    constructor(address admin_, uint delay_) public {
+    constructor(address admin_, uint256 delay_) public {
         require(delay_ >= MINIMUM_DELAY, "Timelock::constructor: Delay must exceed minimum delay.");
         require(delay_ <= MAXIMUM_DELAY, "Timelock::setDelay: Delay must not exceed maximum delay.");
 
@@ -50,7 +50,7 @@ contract CompTimelock {
 
     fallback() external payable {}
 
-    function setDelay(uint delay_) public {
+    function setDelay(uint256 delay_) public {
         require(msg.sender == address(this), "Timelock::setDelay: Call must come from Timelock.");
         require(delay_ >= MINIMUM_DELAY, "Timelock::setDelay: Delay must exceed minimum delay.");
         require(delay_ <= MAXIMUM_DELAY, "Timelock::setDelay: Delay must not exceed maximum delay.");
@@ -76,10 +76,10 @@ contract CompTimelock {
 
     function queueTransaction(
         address target,
-        uint value,
+        uint256 value,
         string memory signature,
         bytes memory data,
-        uint eta
+        uint256 eta
     ) public returns (bytes32) {
         require(msg.sender == admin, "Timelock::queueTransaction: Call must come from admin.");
         require(
@@ -96,10 +96,10 @@ contract CompTimelock {
 
     function cancelTransaction(
         address target,
-        uint value,
+        uint256 value,
         string memory signature,
         bytes memory data,
-        uint eta
+        uint256 eta
     ) public {
         require(msg.sender == admin, "Timelock::cancelTransaction: Call must come from admin.");
 
@@ -111,10 +111,10 @@ contract CompTimelock {
 
     function executeTransaction(
         address target,
-        uint value,
+        uint256 value,
         string memory signature,
         bytes memory data,
-        uint eta
+        uint256 eta
     ) public payable returns (bytes memory) {
         require(msg.sender == admin, "Timelock::executeTransaction: Call must come from admin.");
 
@@ -142,7 +142,7 @@ contract CompTimelock {
         return returnData;
     }
 
-    function getBlockTimestamp() internal view returns (uint) {
+    function getBlockTimestamp() internal view returns (uint256) {
         // solium-disable-next-line security/no-block-members
         return block.timestamp;
     }
